requirements-web.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==4.3.0
PyJWT==2.12.1
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
diskcache==5.6.3
distro==1.9.0
ecdsa==0.19.1
fastapi==0.135.3
gotrue==2.12.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.13.0
langdetect==1.0.9
openai==2.30.0
passlib==1.7.4
postgrest==1.0.1
pydantic-core==2.41.5
pydub==0.25.1
pyasn1==0.6.1
python-dotenv==1.2.2
python-multipart==0.0.22
pyyaml==6.0.3
realtime==2.4.3
requests==2.33.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
storage3==0.11.3
stripe==12.5.0
supabase==2.15.2
supafunc==0.9.4
tenacity==9.1.4
tqdm==4.67.3
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uvicorn==0.43.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
